Initial commit
This is an exciting opportunity for me to get back to my engineering roots and build a website! I expect I'll have tremendous fun tuning this and that. And since I tell engineers that every commit should have a story, I'll do my best to model that. This is the first entry in a log that describes the changes I make to this website, and it has a story.

If you're reading this close to the publication date, you're probably technical. First commits tell a very important part of the development story. I've made a successful living as a software engineer by observing the people and the teams that do it well. I've worked with some of the best, brightest, and most passionate people in the world. And we spent a lot of time figuring out how to build something successful and great:

We talk to users and understand their needs before building, so I will.
We ship frequently, so I will.
We develop in the open, so I will.
We keep things simple, so I will.
We make it secure, so I will.
We make it maintainable, so I will.
We ship it as soon as possible, so I will.
We experiment, so I will.

We know it takes a team, so I'll build one.
We love our country.

Hello America!

This commit serves as the public announcement of my intention to run for the President of the United States of America.

I've spoken with many of my close friends, family, and colleagues about this in varying degrees of detail, and received universal support. I've considered and struggled with it for months - I'll share a lot of my writing from that time.

If everything goes according to the image I have in my head right now, this will start as a personal and campaign website. It will describe my values and my plans. I'll share my pain and my love with the art I create. I'll add data that support my opinions where it matters. It will morph into the official website of the Executive Office of the President of the United States.

Here are some things I love about this approach:

It tracks changes: I'll describe the changes in this log, but you'll be able to see the full history of them over time. I don't think ideas can't change. I don't think people can't change. So my opinions on things are likely to change. And I want to embrace that instead of hide it. I'm not going to stand behind my high school English paper after these two decades! I've learned a lot. But I do want you to read that paper (if someone has it, and I'm hoping they do)! Because knowing what I've learned is also important.

It signs changes: Every change to this website is signed with a special super long password the computer generated for me (RSA). This is extremely hard to guess. It's possible some state actors like China/Russia might be able to do it, so the solution won't work forever, but it will work until an alternative is discussed and decided by someone with more crypto expertise than me. You'll be able to verify (with a higher degree of accuracy than with any other electronic solution) - that I was the source of my words. This is critically important in the world we live in.

Everyone can copy it, fix it, and make updates: This site is available for anybody to copy and modify. I hope people don't do crappy things with this - which is completely possible. But I fully support The Onion or SNL or whoever to use it however they can. I can think of some hilarious variants! I'll put this into the Creative Commons shortly. But you can always verify something is mine by checking the signature. At some point I'd love for everyone in the country to know how to do this verification and these super basic encryption concepts. They're incredibly useful and not really all that hard to figure out. I'll ask the Department of Education what we can do to teach this to our students.

I hope you'll fix a spelling mistake, or suggest changes to data, or make something more beautiful. Maybe even contribute content (though I'm going to be extremely picky about that). Any help at all - I'm sure I'll need it along the way! If someone is interested in contributing but needs assistance with how to work within this system, please email me at jkassemi@gmail.com.